Remove status bits from OpenSSL >=3 version check.
OpenSSL traditionally did not guarantee ABI compatibility across release
(and development) versions.  Because of this, OpenSSH checked the lower 4
"status" bits returned by OpenSSL_version_num(), which were originally
set to 0 for development versions and 0xf for release versions and, if
they did not match, would report the discrepancy and exit.

OpenSSL (unintentionally) changed these bits in the 3.0.0 and subsequent
3.x releases, setting them to zero in the release versions (which happened
to also match the documentation), then changed them back in the 3.5.3
release.  If OpenSSL was upgraded to (or from) this version without
recompiling OpenSSH, it would cause OpenSSH flag it as potentially
incompatible and refuse to use it.  Ultimately OpenSSL rolled this
back, but the check now has no value so is being removed for OpenSSL
versions >=3.

bz#3865 and https://github.com/openssl/openssl/issues/28575, ok djm@