Add sshd_config directive CanonicalMatchUser
Enabling "CanonicalMatchUser" directive in sshd_config will
make sshd to first attempt to obtain canonicalized username
from password database before evaluating the Match User block.
This is mostly useful for AD and LDAP users where using an
alias or just capitalizing some of the letters in a username
still authenticates a user but does not match the Match User
block.

Signed-off-by: Zoltan Fridrich <zfridric@redhat.com>